requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
cloudpickle==2.2.1
cycler==0.11.0
deepdiff==5.8.1
dm-env==1.5
dm-env-rpc==1.1.3
dm-tree==0.1.7
enum34==1.1.10
fonttools==4.34.4
frozenlist==1.3.1
fsspec==2022.8.2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.56.4
grpcio==1.48.1
gym==0.26.2
gym-notices==0.0.8
idna==3.3
immutabledict==2.2.1
importlib-metadata==4.12.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mock==4.0.3
mpyq==0.2.5
multidict==6.0.2
numpy==1.21.5
nvidia-cuda-runtime-cu11==2021.10.25
nvidia-cuda-runtime-cu115==11.5.50
nvidia-pyindex==1.0.9
oauthlib==3.2.1
opencv-python==4.6.0.66
ordered-set==4.1.0
packaging==21.3
pandas==1.5.1
Pillow==9.0.0
pkg_resources==0.0.0
portpicker==1.5.2
protobuf==3.19.4
psutil==5.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pygame==2.1.0
pyparsing==3.0.9
PySC2==4.0.0
python-dateutil==2.8.2
pytorch-lightning==1.7.5
pytz==2022.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s2clientprotocol==5.0.10.88500.0
s2protocol==5.0.10.88500.0
scipy==1.9.1
six==1.16.0
sk-video==1.1.10
SMAC @ git+https://github.com/oxwhirl/smac.git@845b1173c8e38b37f4c67c5a03c6614916fb2f38
tabulate==0.8.9
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torchaudio==0.9.0
torchmetrics==0.9.3
torchvision==0.10.0+cu111
tqdm==4.64.1
typing_extensions==4.0.1
urllib3==1.26.12
websocket-client==1.4.1
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
